+++ tests/ResponseTest.php
@@ -15,7 +15,7 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     use ResponseHelper;
 
     /** @test */
-    public function getDataReturnsAllData()
+    public function it_should_return_data_from_response()
     {
         $response = new Response();
 
@@ -25,7 +25,7 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function getDataReturnsValueByKey()
+    public function it_should_return_value_by_key_from_response()
     {
         $response = new Response();
 
@@ -35,7 +35,7 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function getDataReturnsFalse()
+    public function it_should_return_false_on_error_response()
     {
         $response = new Response();
 
@@ -47,7 +47,7 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function isEmptyReturnsTrueOnEmptyDataOrErrors()
+    public function it_should_check_empty_responses()
     {
         $response = new Response();
 
@@ -59,7 +59,7 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function isEmptyReturnsFalseForData()
+    public function it_should_check_responses_with_data()
     {
         $response = new Response();
 
@@ -69,60 +69,42 @@ class ResponseTest extends PHPUnit_Framework_TestCase
     }
 
     /** @test */
-    public function hasErrorsReturnsTrueForDataWithErrorField()
+    public function it_should_check_responses_with_errors()
     {
         $response = new Response();
 
         $dataWithErrors = $this->createErrorApiResponse('some error');
-
         $this->assertTrue($response->hasErrors($dataWithErrors));
-    }
-
-    /** @test */
-    public function hasErrorsReturnsFalseForData()
-    {
-        $response = new Response();
 
         $data = $this->createSuccessApiResponse();
-
         $this->assertFalse($response->hasErrors($data));
     }
 
+
     /** @test */
-    public function getBookmarksReturnsBookmarksString()
+    public function it_should_return_bookmarks_string_from_response()
     {
         $response = new Response();
 
         $dataWithBookmarks = ['resource' => ['options' => ['bookmarks' => ['my_bookmarks_string']]]];
         $this->assertEquals(['my_bookmarks_string'], $response->getBookmarks($dataWithBookmarks));
-    }
 
-    /** @test */
-    public function getBookmarksReturnsEmptyArrayForNoBookmarks()
-    {
-        $response = new Response();
         $this->assertEmpty($response->getBookmarks([]));
     }
 
     /** @test */
-    public function getPaginationDataReturnsEmptyArrayForNoPagination()
+    public function it_should_return_empty_array_for_response_without_pagination()
     {
         $response = new Response();
         $this->assertEmpty($response->getPaginationData([]));
-    }
-
-    /** @test */
-    public function getPaginationDataReturnsEmptyArrayForErrors()
-    {
-        $response = new Response();
 
         $data = $this->createApiResponse();
-
         $this->assertEmpty($response->getPaginationData($data));
     }
 
+
     /** @test */
-    public function getPaginationDataReturnDataAndBookmarks()
+    public function it_should_return_data_and_bookmarks_from_response_with_pagination()
     {
         $response = new Response();
         $dataWithBookmarks = [